requirements.txt
openai>=1.0.0
vllm>=0.4.0
transformers>=4.40.0
torch>=2.1.0
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
